updated following comments
