--- src/dev/impl/DevToys/ViewModels/Tools/EncodersDecoders/JwtDecoderEncoder/JwtToolJobItem.cs
@@ -1,6 +1,0 @@
-﻿namespace DevToys.ViewModels.Tools.EncodersDecoders.JwtDecoderEncoder
-{
-    public record JwtToolJobItem
-    {
-    }
-}
