+++ src/dev/impl/DevToys/Views/Tools/EncodersDecoders/JwtDecoderEncoder/JwtEncoderControl.xaml.cs
@@ -25,7 +25,6 @@ namespace DevToys.Views.Tools.EncodersDecoders.JwtDecoderEncoder
             set => SetValue(ViewModelProperty, value);
         }
 
-        [ImportingConstructor]
         public JwtEncoderControl()
         {
             InitializeComponent();
